Updated lecture 5
Cleaned up the outline a bit, fixed a name collision issue. (Pandoc
gives each element an id using its text, which was the same as the id I
chose for the `<div>` that hosts the MathScene. Hilarity ensued. Note
for
future: this means we should pay more attention to the scene name!)